Allow common linkage global var have non-zero initializer
In C99 tentative definition of file-scope variable results in common linkage, e.g.

private char *p;
private char *p;

which are initialized by default initializer. For pointer type global variable, the default initializer is null pointer, which does not necessarily have zero value. If a structure contains pointer type member for which null pointer does not have zero value, the strucutre cannot be initialized with zeroinitializer.

Currently IR Verifier asserts when a global variable with common linkage has an initializer which is not zeroinitializer, which it should not. This patch fixes that.

Differential Revision: https://reviews.llvm.org/D27547