Revert "Add FeatureApertureRegs to VI"
This change is incorrect. FeatureApertureRegs is only supported by GFX9.
There is another way to workaround the generic pointer on GPUs not supporting flat
address space. Basically, since flat address space is not supported, we have to
assume that the generic pointer on such GPUs are just place holders. Before they
can be used for load/store, they have to be lowered to real pointers. Since
the lowering has to be done at compile time, it does not matter what is in the
high 32bit of the generic pointer if it is casted from private pointer, so we
can just put 0 there.

This reverts commit e462a1c98d10a4de560e7336fa9cebf86e2c4b76.