Tempororarily disable LowerAlloca pass
In the frontend an addrspacecast is already produced for alloca. Empirical
testing shows adding LowerAlloca would produce additional addrspacecast and
confuses other opt passes, and eventually lead the backend to crash for
ROCm-Device-Libs tests.

This commit disables LowerAlloca pass in AMDGPUTargetMachine, and it allows
ROCm-Device-Libs be built and tested end-to-end.